Check whether `transactions.json` exists before authenticating
Until now, developing the app has been very frustrating because we need
to re-authenticate via email and the app each time we run it. Now that
we're happy that we're retrieving and storing transactions correctly,
it makes sense for us to use local copies of `transactions.json` for
further development, rather than retrieving new transactions each time
we run the script. This commit changes that. Changes include:
- Main script (`launch_dashboard.py`) now checks whether
  `transactions.json` exists before authenticating
- Main script (`launch_dashboard.py`) now only authenticates when we
  need to (i.e. if we want to update `transactions.json` or if it
  doesn't already exist)
- Added a `refresh` function to `src/monzo_api.py` that fetches all
  transactions from some specified date `last_transaction`.
- Moved some authentication code from `launch_dashboard.py` to
  `src/auth.py` to make the former more readable.