Only force re-authentication where necessary
There's no need to force the user to re-authenticate if they don't want
to interact with Monzo's servers. Now re-authentication is only forced
if the user presses the "Update transactions" button when their
`access_token` has expired. Previously, re-authentication would be
forced even if the user was just running the app locally.

To achieve this, I added an `access_token_refresh_needed` function to
`src/utils.py` that can be called in route handlers that deal with API
requests. Currently, this is just called by the `/update-transactions`
route handler in `src/app.py`.